Update README.md
This repository is aimed at the problems, which I encounter from time to time. You won't find the tool here. You can see what issue I have at the given time and try to help me if you can. It will be cordially appreciated. We all know, that solving some problem in public we are giving a chance to learn for anybody, who has the same or similar issue.